Change plugin API: initialize plugins in their load() functions
This makes the whole plugin initialization more flexible; plugins are
free to do anything with the bot object. However, it might be a good
idea to pass a specially crafted view/context/interface instead, but
right now I don't what needs to be exposed so I decided to pass the
complete bot object.

It goes without saying that the plugin API will have backward
incompatible changes in the future, but hey, we are in the development
phase right now.