core: do not delete plugin modules from sys.modules
Plugin modules might import something and if the plugin module is
deleted then the module imported by the plugin module is not available
in the plugin object's context anymore. For example plugins/rcon import
subprocess which is referenced in _RconPlugin.__crcon() method. If rcon
module is deleted then the reference to the subprocess is lost and
subprocess is None.